+++ gdsfactory/components/C.py
@@ -11,7 +11,9 @@ def C(
     size: Tuple[float, float] = (10.0, 20.0),
     layer: LayerSpec = "WG",
 ) -> Component:
-    """Generates a 'C' geometry with ports on both ends. Adapted from phidl
+    """C geometry with ports on both ends.
+
+    Adapted from phidl.
 
     Args:
         width: of the line.
@@ -19,6 +21,7 @@ def C(
         layer: layer spec.
 
     .. code::
+
          ______
         |       o1
         |   ___
@@ -44,8 +47,8 @@ def C(
         (-w, -w),
     ]
     c.add_polygon(points, layer=layer)
-    c.add_port(name="o1", midpoint=(s1, s2), width=width, orientation=0, layer=layer)
-    c.add_port(name="o2", midpoint=(s1, 0), width=width, orientation=0, layer=layer)
+    c.add_port(name="o1", center=(s1, s2), width=width, orientation=0, layer=layer)
+    c.add_port(name="o2", center=(s1, 0), width=width, orientation=0, layer=layer)
     return c