+++ gdsfactory/components/mmi2x2.py
@@ -75,23 +75,19 @@ def mmi2x2(
     )
 
     ports = [
-        gf.Port(
-            "o1", orientation=180, midpoint=(0, -a), width=w_taper, cross_section=x
-        ),
-        gf.Port(
-            "o2", orientation=180, midpoint=(0, +a), width=w_taper, cross_section=x
-        ),
+        gf.Port("o1", orientation=180, center=(0, -a), width=w_taper, cross_section=x),
+        gf.Port("o2", orientation=180, center=(0, +a), width=w_taper, cross_section=x),
         gf.Port(
             "o3",
             orientation=0,
-            midpoint=(length_mmi, +a),
+            center=(length_mmi, +a),
             width=w_taper,
             cross_section=x,
         ),
         gf.Port(
             "o4",
             orientation=0,
-            midpoint=(length_mmi, -a),
+            center=(length_mmi, -a),
             width=w_taper,
             cross_section=x,
         ),