+++ gdsfactory/samples/11_component_layout.py
@@ -50,11 +50,11 @@ def test_crossing_arm(
     c.add_polygon(taper_points, layer=LAYER.WG)
 
     c.add_port(
-        name="o1", midpoint=(-xmax, 0), orientation=180, width=wg_width, layer=LAYER.WG
+        name="o1", center=(-xmax, 0), orientation=180, width=wg_width, layer=LAYER.WG
     )
 
     c.add_port(
-        name="o2", midpoint=(xmax, 0), orientation=0, width=wg_width, layer=LAYER.WG
+        name="o2", center=(xmax, 0), orientation=0, width=wg_width, layer=LAYER.WG
     )
     return c